Move CustomColors, CustomIcons, DBHelper in utils
