+++ lib/card_items/transaction_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../custom_colors.dart';
+import '../utils/custom_colors.dart';
 import '../models/transaction.dart';
 
 // Used in HistoryScreen.
