+++ lib/screens/edit_transaction_screen.dart
@@ -2,7 +2,7 @@ import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter/material.dart';
 
-import '../custom_colors.dart';
+import '../utils/custom_colors.dart';
 import '../card_items/label_chooser_card.dart';
 import '../models/transaction.dart';
 import '../models/label.dart';
