+++ lib/widgets/transaction_details_appbar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../custom_colors.dart';
+import '../utils/custom_colors.dart';
 
 // This is used in TransactionDetailsScreen.
